''Trunk Update'' '''Type:''' New feature: technical '''Ticket:''' #53
__Overview__
A runtime checking mechanism (at least for offline/serial apps anyway), to verify that technical changes and/or science changes behind switches that when they are turned off do NOT change the model output.

Please Note: This tool is restricted to serial, offline applications. The ticket will be re-opened or a new ticket generated to implement in the UM and/or for multiple processors.

__Technical comments__
_driver sums balances and writes to a hidden file. This summed balance is compared within the model to the summed balance written to a hidden file generated by running an earlier, standard revision. Typically this standard revision will be the point on the trunk from which you branched.

See Ticket #53 for further details on usage.

__Science comments__
neutral


git-svn-id: https://trac.nci.org.au/svn/cable/trunk@2285 b483b3d9-cc7d-4f55-9e4c-b4562452379e