''Trunk Update'' '''Type:''' Technical correction. '''Ticket:''' #101
__Overview__
There are still issues surrounding the implementation of Ticket #2 and the introduction of existing variables to be initialized through the def_veg_params file instead of being hard-wired. This seems to be an issue only with the online/ACCESS runs. In this instance it was noted that the GPP was not right.

We have modified code here to utilize the same code in online and offline modes. Whilst this seems to fix the GPP issue, there remains an issue of freproducability between runs online

__Technical comments__
Ticket #2 introduced the initialization of existing variables through the def_veg_params file rather than being hard wired. There have been issues of this affecting the online/ACCESS runs, whereas there. In this instance it was noted that the GPP was still nor right. 

We have modified code here to utilize the same code for reading and initializing veg% variables in online and offline modes. Whilst this seems to fix the GPP issue, there remains an issue of freproducability between runs online

__Science comments__
It was noted by Tilo that GPP was not right in post Ticket #2 revisions, and traceable to initialization of those implemented variables.




git-svn-id: https://trac.nci.org.au/svn/cable/trunk@3634 b483b3d9-cc7d-4f55-9e4c-b4562452379e