''Trunk Update'' '''Type:''' Corrected defect. '''Ticket:''' #73
__Overview__
Initialize kwidth_gl in offline simmulations. 
kwidth_gl is used in updating snow_age in cable_albedo.

__Technical comments__
Minor change. USE kwidth_gl element of _common module in drivers and initialize.

__Science comments__
cable_albedo uses dtau to increment the snow age where appropriate. Calculation of dtau uses the width of the timestep. Online this is used as kwidth_gl. Offline this was not being initialized. Now Corrected.  
NOT science neutral but 1. only affects snow covered tiles and  2, difference is not earth shattering.



git-svn-id: https://trac.nci.org.au/svn/cable/trunk@3806 b483b3d9-cc7d-4f55-9e4c-b4562452379e