''Trunk Update''
'''Type:'''
Bug Fix

'''Ticket:'''
#120

__Overview__
There appears to have been a long-standing logic error in the nmeth=1 method used to compute soil moisture diffusion through layers. HOWEVER - this method is not used anyway. I had to force it to gauge the impact of the modifications. This is linked in comment 8 of the ticket #120.



git-svn-id: https://trac.nci.org.au/svn/cable/trunk@8660 b483b3d9-cc7d-4f55-9e4c-b4562452379e