''Trunk Update''
'''Type:'''
Give default value to gs_switch

'''Ticket:'''
#332

__Overview__
CABLE won't run without a value for gs_switch being defined. The intention was to force the user to specify a stomatal conductance scheme, however this is reported as a problem often enough that is preferable to define a default scheme here. This was intended to upload as aprt of #332, however the bulk of #332 has been put on hold until a thourough assesment can be achieved


git-svn-id: https://trac.nci.org.au/svn/cable/trunk@9000 b483b3d9-cc7d-4f55-9e4c-b4562452379e