Deprecate `serial_cable` and `parallel_cable` scripts
This change simplifies the CABLE offline build system by deprecating the
`serial_cable` and `parallel_cable` build scripts and moving their
functionality into the Makefile.

Replace `ls *.o` in `serial_cable` and `parallel_cable` scripts with
`$(OBJS)` in the Makefile where `OBJS` is a list of object files common
to both serial and MPI. This is done so that:

1. We leverage `make` syntax to simplify the rules for `serial` and
   `mpi` targets.
2. We can replace the `all` target (which contains a list of object
   files as dependencies) with `$(OBJS)`.
3. We clean up the existing definition of `LSRC` in the Makefile and use
   this to generate a list of the required object files.

See [#193][PR] for a sorted diff of the new `LSRC` variable against the
previous `LSRC` and against the list of object file dependencies in the
`all` target.

Tools for testing binary equivalence of executables (see `ldd`, `nm` and
`objdump` tools), show that the serial and MPI executables are
equivalent to their respective executable in the main branch.

Fixes #189

[PR]: https://github.com/CABLE-LSM/CABLE/pull/193