Calendar attribute independent from leap years (#195)
# CABLE

Thank you for submitting a pull request to the CABLE Project.

## Description

Removes the dependency of the calendar attribute on the fact we are
running a leap year or not. So output files from a same configuration
will have the same calendar attribute for all years.
Fixes #187 

## Testing
**Setup:** ran a global simulation with CRU JRA forcing (using gswp3
namelist settings) with both this branch and the main branch for 2
years: normal and leap year.

**Results:** `cdo diffn` indicates the variables in the cable_out.nc
files are identical. But the calendar attribute for `time` is different
in the non-leap year: `noleap` with the code from the main branch,
`standard` with the code from this branch.

## Type of change

Please delete options that are not relevant.

- [X] Bug fix

## Documentation

No changes to the documentation are required for this change.

Please add a reviewer when ready for review.


<!-- readthedocs-preview cable start -->
----
📚 Documentation preview 📚:
https://cable--195.org.readthedocs.build/en/195/

<!-- readthedocs-preview cable end -->

---------

Co-authored-by: ccc561 <ccc561@b483b3d9-cc7d-4f55-9e4c-b4562452379e>
Co-authored-by: C. Carouge <c.carouge@unsw.edu.au>