Remove `-O0` flag from `LDFLAGS`
In `build3.sh`, the `-O0` flag (which disables optimisations) is
included in the `LDFLAGS` variable. This change removes the `-O0` flag
from `LDFLAGS` so that:

1. Compiler specific flags (non `ld` flags) are removed from `LDFLAGS`
2. `-O0` flag is not applied for only a subset of source files.
3. The Makefile rules can be simplified.

See #196 for more details on why the flag should be removed.

Regression tests (bitwise comparison of model output via `nccmp`) show
that model output is bitwise identical between the current branch and
the main branch for serial and MPI model runs.

Fixes #196