Add support for MPI builds with the GNU compiler (#241)
The build script currently does not support MPI builds with the GNU
compiler on Gadi. This is due to the `intel-mpi` module being loaded
when we should be using an MPI module that is compatible with the chosen
compiler. This change loads the `openmpi` module when GNU compilation is
enabled in the build script.

Fixes #239

<!-- readthedocs-preview cable start -->
----
📚 Documentation preview 📚:
https://cable--241.org.readthedocs.build/en/241/

<!-- readthedocs-preview cable end -->