CABLE3 version of ruff_resist/init_radiation is implemented and the corresponding changes to CALLed subroutines and CALLing cable_cbm. Necessary changes for it to build have been made in the code and building scripts. pop_fbeam switch has also been created to switch back to POP behavior. i.e. with this switch = .TRUE. CABLE-POP-TRENDY behaviour is returned cable_air, cable_carbon and a few others have also been merged with CABLE3 versions
diff -q -r CABLE3 this branch' will show the oustanding files that are different. cbl_init_radiation for example will show up as one of these files even though it has been merged. Closer inspection of the actal differences with vimdiff or whatever will show what is actally different AND significantly these differences are behind a switch so that we can toggle between behavioursdiff or whatever will show what is actally different AND significantly these differences are behind a switch so that we can toggle between behaviours.

Next we will add the merged albedo/ directory.