Large set of changes generalising the input routines for the Met forcing.
These changes permit the use of input datasets spanning multiple files by specifying a template which matches all the files in the dataset. The files in the dataset must contain
YYYYMMDD specifications for the start and end date of the data contained in the specific file. The entry in the namelist is the template of the dataset, with <startdate> and <enddate> replacing the start and end date in the filename.