Reconcile versions of rplant (#274)
Fixes #163 

initialization to 1 so is applicable even IF not over-written by
specific case applications. i.e. Will be correct for ESM1.5

# CABLE

Thank you for submitting a pull request to the CABLE Project.

## Description

Please include a brief summary of the change and list the issues that
are fixed.
Please also include relevant motivation and context.

You can link issues by using a supported keyword in the pull request's
description or in a commit message:

Fixes #(issue)

## Type of change

Please delete options that are not relevant.

- [x] Bug fix
- [ ] New or updated documentation

## Checklist

- [x] The new content is accessible and located in the appropriate
section.
- [x] I have checked that links are valid and point to the intended
content.
- [x] I have checked my code/text and corrected any misspellings

Please add a reviewer when ready for review.


<!-- readthedocs-preview cable start -->
----
📚 Documentation preview 📚:
https://cable--274.org.readthedocs.build/en/274/

<!-- readthedocs-preview cable end -->