Add routine to search for dimension IDs by name. (#377)
# CABLE

Thank you for submitting a pull request to the CABLE Project.

## Description

The retrieving of dimension IDs from input NetCDF files is inconsistent
across the code. Add routines to the ```cable_common``` module to
standardise the retrieving of said IDs by searching through a list of
possible names for the dimension.

Fixes #(issue)

## Type of change

- [x] Feature

## Checklist

- [x] The new content is accessible and located in the appropriate
section.
- [x] I have checked that links are valid and point to the intended
content.
- [x] I have checked my code/text and corrected any misspellings

Please add a reviewer when ready for review.