Add a new module to handle MPI, including: - routines to initialise and finalise MPI - a simple class to handle MPI groups.
All the contents of the module can handle the case where no MPI support is available and/or serial runs (i.e., use of only one rank), thus removing the need to fence some operations with preprocessing directives outside of this module.

To get CMake to properly propagate the link dependencies, I had to convert the common object library into a static library.