431 add routine for more informative namelist errors (#438)
# CABLE

Thank you for submitting a pull request to the CABLE Project.

## Description

Add a routine to handle namelist errors, that returns a much more
informative error message than the baseline namelist error message.
Utilises the new ```IOMSG``` specifier and the routine detailed
[here](https://degenerateconic.com/namelist-error-checking.html).

Fixes #(431)

Edit: Turns out that this namelist checking routine relies on undefined
behaviour. The problems I was seeing was the result of a "regression" in
Intel Fortran 2023, see [this
thread](https://community.intel.com/t5/Intel-Fortran-Compiler/Regression-with-quot-backspace-quot-in-2023-0/m-p/1448175).
The comment from Steve Lionel in the accepted answer is important:

> The standard says, "If an error condition occurs during execution of
an input/output statement, the position of the file becomes
indeterminate."

This means that the so-called regression was not really a regression,
just a change in already undefined behaviour. This use of
```BACKSPACE``` to return the incriminating line relies on undefined
behaviour that may change at any time. Based on this, I decided against
including this part of the routine, and just throwing back the
```IOMSG``` returned by the ```READ``` call.

## Type of change

- Quality of life change