Add Fortran2008 standards checker for Intel and GNU builds (#470)
This flag just enables warnings when code is not Fortran2008 compliant.
Should have 0 performance effect. Currently added to all builds, might
move to only DEBUG builds.

# CABLE

Thank you for submitting a pull request to the CABLE Project.

## Description

Just adds standard checking flags to the Intel and GNU builds (```-stand
f08``` and ```std=f08``` respectively). This throws warnings when
something is not compliant with the Fortran 2008 standards, and should
have no effect on performance. Currently added the flag to all builds,
but maybe it should be moved to ```DEBUG``` builds only?

Doesn't fix, but part of the wider issue around #449 

## Type of change

- Code clean-up

## Checklist

- [x] The new content is accessible and located in the appropriate
section.
- [x] I have checked that links are valid and point to the intended
content.
- [x] I have checked my code/text and corrected any misspellings

Please add a reviewer when ready for review.