Move LALLOC initialisation into cable_driver_init
Note LALLOC is not added as a namelist switch for this change as this
should be acompanied with a description of the behaviour of LALLOC
(which I am not familiar with). However, LALLOC is declared as a module
variable so that it can be easily added as a namelist switch.