Use NEWUNIT in almost all active OPEN statements (#460)
# CABLE

Thank you for submitting a pull request to the CABLE Project.

## Description

Extend the use of the ```NEWUNIT=``` specifier for acquiring io units to
the rest of the active ```OPEN``` calls in the code. There is one
exception, in the calls found within
```core/biogeophys/cable_diag.F90```. The units are handled a bit
strangely here, with one being defined dependent on the other. Looks
like it might interact with the UM coupling in a specific way, so I've
left it as is at the moment.

Fixes #455 

## Type of change

- Fortran standards update

## Checklist

- [ ] The new content is accessible and located in the appropriate
section.
- [ ] I have checked that links are valid and point to the intended
content.
- [ ] I have checked my code/text and corrected any misspellings