changes required for ESM16 (#474)
# CABLE

Thank you for submitting a pull request to the CABLE Project.

## Description

With these minor changes CABLE:main can be cloned directly into the
ESM1.6 / UM7.3 code under umbase_hg3. NB:A lot of the changes are
replicated from PRs that are waiting to be reviewed.

Fixes #473, #468

## Type of change

For ESM1.6. A library can be built directly from this code

benchcab regression tessts successful:  

[benchmark_cable_qsub.sh.o128621374.txt](https://github.com/user-attachments/files/17748080/benchmark_cable_qsub.sh.o128621374.txt)




<!-- readthedocs-preview cable start -->
----
📚 Documentation preview 📚:
https://cable--474.org.readthedocs.build/en/474/

<!-- readthedocs-preview cable end -->

---------

Co-authored-by: Claire Carouge <claire.carouge@anu.edu.au>