494 redux local ice fixes (#527)
# CABLE

Thank you for submitting a pull request to the CABLE Project.

## Description

Please include a brief summary of the change and list the issues that
are fixed.
Please also include relevant motivation and context.

You can link issues by using a supported keyword in the pull request's
description or in a commit message:

Fixes #(issue)

## Type of change

Please delete options that are not relevant.

- [ ] Bug fix
- [ ] New or updated documentation

## Checklist

- [ ] The new content is accessible and located in the appropriate
section
- [ ] I have checked that links are valid and point to the intended
content
- [ ] I have checked my code/text and corrected any misspellings

## Testing

- [ ] Are the changes bitwise-compatible with the main branch? If
working on an optional feature, are the results bitwise-compatible when
this feature is off? If yes, copy benchcab output showing successful
completion of the bitwise compatibility tests or equivalent results
below this line.

- [ ] Are the changes non bitwise-compatible with the main branch
because of a bug fix or a feature being newly implemented or improved?
If yes, add the link to the modelevaluation.org analysis versus the main
branch or equivalent results below this line.

Please add a reviewer when ready for review.


<!-- readthedocs-preview cable start -->
----
📚 Documentation preview 📚:
https://cable--527.org.readthedocs.build/en/527/

<!-- readthedocs-preview cable end -->

---------

Co-authored-by: Claire Carouge <claire.carouge@anu.edu.au>