Avoid using uninitialised trunk_sumbal
The trunk_sumbal value is still used in the offline driver when a
consistency check file cannot be find. To avoid using a garbage value
and potentially causing an exception, this change moves the tolerance
calculation and the trunk_sumbal initialisation into a single subroutine
so that trunk_sumbal is used only when it is successfully read from the
the consistency check file.