Avoid using uninitialised trunk_sumbal
The trunk_sumbal value is still used in the offline driver when a
consistency check file cannot be find. To avoid using a garbage value
and potentially causing an exception, this change promotes trunk_sumbal
to an allocatable so that the variable is used only when it is
allocated and initialised via the consistency check file.