following #220 epot calc in AM3 (#555)
# CABLE

Thank you for submitting a pull request to the CABLE Project.

## Description

[See](https://github.com/CABLE-LSM/ACCESS3-JULES/pull/250) 

Fixes #554 

## Type of change

- [ ] Bug fix

## Checklist

- [x] The new content is accessible and located in the appropriate
section
- [x] I have checked that links are valid and point to the intended
content
- [x] I have checked my code/text and corrected any misspellings

## Testing

- [x] Are the changes bitwise-compatible with the main branch? If
working on an optional feature, are the results bitwise-compatible when
this feature is off? If yes, copy benchcab output showing successful
completion of the bitwise compatibility tests or equivalent results
below this line.

- [ ] Are the changes non bitwise-compatible with the main branch
because of a bug fix or a feature being newly implemented or improved?
If yes, add the link to the modelevaluation.org analysis versus the main
branch or equivalent results below this line.

**2025-02-28 18:28:43,893 - INFO - benchcab.benchcab.py:391 - 0 failed,
168 passed**

<!-- readthedocs-preview cable start -->
----
📚 Documentation preview 📚:
https://cable--555.org.readthedocs.build/en/555/

<!-- readthedocs-preview cable end -->