reconcile with dev-access-esm1.6 - interface code (#551)
# CABLE

## Description
Reconcile developments in ESM1.6 with CABLE:main

There are changes across utilities since directories as well BUT here it
is only those in coupled/ that are considered.

Post shuffling these files around to facilitate building a library. What
happens to them is irrelevant really.

These modified and added files are from dev-access-esm1.6. 

## Testing

- [x] Are the changes bitwise-compatible with the main branch? If
working on an optional feature, are the results bitwise-compatible when
this feature is off? If yes, copy benchcab output showing successful
completion of the bitwise compatibility tests or equivalent results
below this line.

So - I haven't actually done the testing BUT considering none of these
fits are even in the build they cant impact anything

<!-- readthedocs-preview cable start -->
----
📚 Documentation preview 📚:
https://cable--551.org.readthedocs.build/en/551/

<!-- readthedocs-preview cable end -->