Add namelist option for CASA output file path (#563)
Currently the CASA output file is set by the `filename%path`,
`cable_user%RunIden`, `cable_user%YEARSTART` and `cable_user%YEAREND`
namelist options and is not easily configurable. This change adds an
optional namelist option `casafile%out` (as per the CABLE-POP_TRENDY
branch) so that the CASA output path can be set explicitly from the
namelist file.

## Type of change

- [x] New or updated documentation

## Checklist

- [x] The new content is accessible and located in the appropriate
section
- [x] I have checked that links are valid and point to the intended
content
- [x] I have checked my code/text and corrected any misspellings

## Testing

- [x] Are the changes bitwise-compatible with the main branch? If
working on an optional feature, are the results bitwise-compatible when
this feature is off? If yes, copy benchcab output showing successful
completion of the bitwise compatibility tests or equivalent results
below this line.

```
2025-03-12 16:36:52,612 - INFO - benchcab.benchcab.py:381 - tasks: 168 (models: 2, sites: 42, science configurations: 4)
2025-03-12 16:39:50,137 - INFO - benchcab.benchcab.py:391 - 0 failed, 168 passed
```

<!-- readthedocs-preview cable start -->
----
📚 Documentation preview 📚:
https://cable--563.org.readthedocs.build/en/563/

<!-- readthedocs-preview cable end -->