Avoid using uninitialised trunk_sumbal (#568)
The trunk_sumbal value is still used in the offline driver when a
consistency check file cannot be find. To avoid using a garbage value
and potentially causing an exception, this change moves the tolerance
calculation and the trunk_sumbal initialisation into a single subroutine
so that trunk_sumbal is used only when it is successfully read from the
the consistency check file.

## Type of change

Please delete options that are not relevant.

- [x] Bug fix

## Checklist

- [x] The new content is accessible and located in the appropriate
section
- [x] I have checked that links are valid and point to the intended
content
- [x] I have checked my code/text and corrected any misspellings

## Testing

- [x] Are the changes bitwise-compatible with the main branch? If
working on an optional feature, are the results bitwise-compatible when
this feature is off? If yes, copy benchcab output showing successful
completion of the bitwise compatibility tests or equivalent results
below this line.

```
2025-03-25 11:11:26,592 - INFO - benchcab.benchcab.py:380 - Running comparison tasks...
2025-03-25 11:11:26,627 - INFO - benchcab.benchcab.py:381 - tasks: 168 (models: 2, sites: 42, science configurations: 4)
2025-03-25 11:14:17,701 - INFO - benchcab.benchcab.py:391 - 0 failed, 168 passed
```

<!-- readthedocs-preview cable start -->
----
📚 Documentation preview 📚:
https://cable--568.org.readthedocs.build/en/568/

<!-- readthedocs-preview cable end -->