Fix library build for integration into UM7 (#590)
# CABLE

Thank you for submitting a pull request to the CABLE Project.

## Description

Fixes the library build after the re-shuffle to align the coupled and
offline versions of the code. Some files are removed from the library
build, some are pulled from the coupled version within UM7 and added to
the library build. Brings the recent changes post-shuffle from Jhan's
work in [his
branch](https://github.com/CABLE-LSM/CABLE/tree/reconcile_local_May5) to
include australian PFTs. The code builds in offline mode, and UM7 has
been successfully compiled using the generated library.

Testing for bitwise equivalence of ESM1.6 with the library are ongoing.
Yet to run through Benchcab.

## Testing

- [ ] Are the changes bitwise-compatible with the main branch? If
working on an optional feature, are the results bitwise-compatible when
this feature is off? If yes, copy benchcab output showing successful
completion of the bitwise compatibility tests or equivalent results
below this line.

- [ ] Are the changes non bitwise-compatible with the main branch
because of a bug fix or a feature being newly implemented or improved?
If yes, add the link to the modelevaluation.org analysis versus the main
branch or equivalent results below this line.


<!-- readthedocs-preview cable start -->
----
📚 Documentation preview 📚:
https://cable--590.org.readthedocs.build/en/590/

<!-- readthedocs-preview cable end -->

---------

Co-authored-by: Srbinovsky, Jhan (O&A, Aspendale) <Jhan.Srbinovsky@csiro.au>
Co-authored-by: Claire Carouge <claire.carouge@anu.edu.au>