Make the code compute the diffuse fraction by default (#603)
# CABLE

## Description

The ```CABLE-POP_TRENDY``` branch of the code currently sets
```cable_user%calc_fdiff=.FALSE.``` by default. The only place this was
modified was in the CRU meteorology setup, so would effectively break
anyone else's simulation not using this workflow who was unaware of this
change.

This change changes the default to ```cable_user%calc_fdiff=.TRUE.```,
so that the code computes the direct and diffuse fractions via the
spitter function. The check used in the CRU/TRENDY has been moved to
initialisation. If the user turns off diffuse fraction calculation (via
the ```ReadDiffFrac=.TRUE.``` option in the CRU namelist), but doesn't
also provide a dataset to read from, the code will now throw an error.

This is not an ideal fix, as we should be checking against the
```cable_user%calc_fdiff``` option, rather than the
```CRU%ReadDiffFrac``` option. I think a better option is to remove the
```CRU%ReadDiffFrac``` option altogether and only have the
```cable_user%calc_fdiff```, given they effectively contain identical
information. That would also require minor adjustments to the TRENDY
scripts.

Fixes at least part of the problems discussed in #526.

## Type of change

- [x] Bug fix