Use mpi_f08 in `cable_mpi.F90` (#616)
In our mock CABLE prototype, @micaeljtoliveira, @Whyborn and I made the
change to use the Fortran 2008 MPI types and procedures via the
`mpi_f08` module. The new interfaces provide type safety by
distinguishing various MPI types, instead of using integer for all
types, as is done in the old interfaces.

This feature has been available already for a few years:
- OpenMPI >= 1.7 [2013]
  https://github.com/open-mpi/ompi/blob/v1.7.0/NEWS#L65
- MPICH >= 3.1.1 [2014]
  https://github.com/pmodels/mpich/blob/v3.1.1/CHANGES#L10-L11
- Intel MPI >= 5.1 (needs Intel Fortran Compiler >= 16.0, part of Intel
Parallel Studio XE 2016) [2015]

https://www.intel.com/content/dam/develop/external/us/en/documents/intelmpi-5-1-update3-build223-releasenotes-linux.pdf

This change brings across these changes to CABLE main. The new interface
will be required for the new MPI implementation (#358).

## Type of change

Please delete options that are not relevant.

- [x] New feature

## Checklist

- [x] The new content is accessible and located in the appropriate
section
- [x] I have checked that links are valid and point to the intended
content
- [x] I have checked my code/text and corrected any misspellings

## Testing

- [x] Are the changes bitwise-compatible with the main branch? If
working on an optional feature, are the results bitwise-compatible when
this feature is off? If yes, copy benchcab output showing successful
completion of the bitwise compatibility tests or equivalent results
below this line.
```
2025-10-03 12:21:47,177 - INFO - benchcab.benchcab.py:379 - Running comparison tasks...
2025-10-03 12:21:47,204 - INFO - benchcab.benchcab.py:380 - tasks: 168 (models: 2, sites: 42, science configurations: 4)
2025-10-03 12:24:39,211 - INFO - benchcab.benchcab.py:390 - 0 failed, 168 passed
```

<!-- readthedocs-preview cable start -->
----
📚 Documentation preview 📚:
https://cable--616.org.readthedocs.build/en/616/

<!-- readthedocs-preview cable end -->