Cable updates from esm16 merge1and2 (#626)
# CABLE

Thank you for submitting a pull request to the CABLE Project.

## Description

It might be beneficial to have these all in one place. Given a few
trivial changes that really should go into the ESM side, and retained
here (e.g referencing cable_serial replacing cable_driver) - comparing
this branch to the [ESM1.6
branch](https://github.com/ACCESS-NRI/UM7/pull/170) there are ZERO
differences.

merging branches representative of #622 #623 #624


Fixes #619 #628

## Type of change

Please delete options that are not relevant.

- [NA] Bug fix
- [NA] New or updated documentation

## Checklist

- [x] The new content is accessible and located in the appropriate
section
- [x] I have checked that links are valid and point to the intended
content
- [x] I have checked my code/text and corrected any misspellings

## Testing

- [ ] Are the changes bitwise-compatible with the main branch? If
working on an optional feature, are the results bitwise-compatible when
this feature is off? If yes, copy benchcab output showing successful
completion of the bitwise compatibility tests or equivalent results
below this line.

- [ ] Are the changes non bitwise-compatible with the main branch
because of a bug fix or a feature being newly implemented or improved?
If yes, add the link to the modelevaluation.org analysis versus the main
branch or equivalent results below this line.

Please add a reviewer when ready for review.


<!-- readthedocs-preview cable start -->
----
📚 Documentation preview 📚:
https://cable--626.org.readthedocs.build/en/626/

<!-- readthedocs-preview cable end -->