meaningful description of switch (#634)
# CABLE

Thank you for submitting a pull request to the CABLE Project.

## Description

The history of this switch was that in ACCESS1.3, when we first
introduced 17 tiles, and tiled soils, manipulation of fields files was
way harder than what it is now. Plus, we didn't have a similar field to
use as a base. We began with files of constant temperature, moisture
etc. Progressively made amendments. In one iteration (lasting months
perhaps), we clobbered values on snow fields in particular by calling a
special initialization routine. This was more or less the first time was
coupled (to the UM) - hence the naming of the flag as runtime_coupled.
As time went on we occasionally used it in other scenarios. The meaning
of the flag and its declaration initialization got mixed up until
eventually it was determined that while the flag was TRUE, the
conditions should be IF .NOT. . It is all very convoluted evolution to
basically set a switch that triggers a special (albeit basic)
initialization of soilnow fields in the instance where the
initializations through a restart (etc) cannot be trusted.

It make way more sense to rename the switch, declare it as FALSE, and
update the conditionals rather than than have this crazy double negative
situation.

Fixes #536 

## Type of change

CLARIFICATION

## Checklist

- [x] The new content is accessible and located in the appropriate
section
- [x] I have checked my code/text and corrected any misspellings

## Testing

- [x] Are the changes bitwise-compatible with the main branch? If
working on an optional feature, are the results bitwise-compatible when
this feature is off? If yes, copy benchcab output showing successful
completion of the bitwise compatibility tests or equivalent results
below this line.

<!-- readthedocs-preview cable start -->
----
📚 Documentation preview 📚:
https://cable--634.org.readthedocs.build/en/634/

<!-- readthedocs-preview cable end -->