Introduce MPI-aware structure variables
Sub-dividing the grid based on the number of MPI ranks available is key
to enabling parallelism in CABLE. This change achieves this by
introducing several new local structure variables to describe the grid
information local to the current MPI rank. These are `mp`, `mland`,
`land_x`, `land_y`, `land_decomp_start`, `land_decomp_end`,
`patch_decomp_start`, `patch_decomp_end` and `landpt`. Global structure
variables are also introduced to describe the grid of the global
simulation domain, irrespective of the current MPI rank, as this
information is necessary for deriving several local structure variables
and for generating parallel I/O decompositions.  These are
`mland_global`, `land_x_global`, `land_y_global`, `mp_global` and
`landpt_global`.

This change also updates constant arrays which are initialised in
`open_met_file` to be stored on the global land grid on each MPI rank.
This is required for correctly initialising global structure variables
such as `landpt_global` which is dependent on the latitude and longitude
coordinates of each grid cell in the global simulation domain. These
variables are:

- `cable_iovars.F90:latitude(mland_global)`
- `cable_iovars.F90:longitude(mland_global)`
- `cable_iovars.F90:vegtype_metfile(mland_global, nmetpatches)`
- `cable_iovars.F90:vegpatch_metfile(mland_global, nmetpatches)`
- `cable_iovars.F90:soiltype_metfile(mland_global, nmetpatches)`
- `cable_iovars.F90:PrecipScale(mland_global)`
- `cable_input.F90:elevation(mland_global)`
- `cable_input.F90:avPrecip(mland_global)`

Note that for the legacy MPI case, the MPI group used to initialise the
structure variables on the master driver will always be a single rank
group as the master driver will only perform serial I/O.