`build-ci`: Pin `spack-config-ref` to fix breaking `spack-config` change (#653)
# CABLE

Thank you for submitting a pull request to the CABLE Project.

## Description

We are updating our `build-ci` infrastructure so it will be able to
support `spack >= v1.0` more fully. There is an incoming change (see
https://github.com/ACCESS-NRI/spack-config/pull/82) to `spack-config` to
forward this goal, that will no longer work with `build-ci@v2`. There is
a new version of `build-ci@v3` (see
https://github.com/ACCESS-NRI/build-ci/pull/253) that will work with
this new version of `spack-config`, but it is not yet ready for merging.

This interim fix for `build-ci@v2` will work while `build-ci@v3` is
being tested, so there is no downtime between versions, by pinning the
version of `spack-config` to a tag before the breaking change.

> [!NOTE]
> Open pull requests will need to be rebased to get this fix, or a
similar pinning of `spack-config-ref` in the PR added

## Type of change

- [x] Bug fix


<!-- readthedocs-preview cable start -->
----
📚 Documentation preview 📚:
https://cable--653.org.readthedocs.build/en/653/

<!-- readthedocs-preview cable end -->