Move `range_abort` from cable_abort.F90 to cable_checks.F90
This change allows the `cable_abort_module` to be used in modules where
`cable_def_types_mod` or `cable_IO_vars_module` are a dependee of that
module as the removal of `range_abort` avoids introducing cyclic module
dependencies. The impact of this change is minimal as `range_abort` is
only called from the `cable_abort_module` in the code base.